move qrScanService to page/qr_scan directory
